kickstart.tmpl: stop using depricated "installerfw" keyword
We renamed the "installerfw" installer framework keyword to
"installerfw_plugins" in MIC.  MIC still supports both, but the intention is to
remove the old one.

Lets start using "installerfw_plugins" keyword from now on.

This should fix the following MIC warning:

"please, use 'installerfw_plugins' instead of 'installerfw', the latter is
obsolete and will be removed in future releases"

Change-Id: I21911db1a17581a8afdab934c7f7fb0a2e1852c9
Signed-off-by: Artem Bityutskiy <artem.bityutskiy@linux.intel.com>